backend/requirements.txt
anicli-ru
pysocks
pyinstaller